+++ mayan/apps/linking/links.py
@@ -50,8 +50,9 @@ link_smart_link_instance_view = Link(
     )
 )
 link_smart_link_instances_for_document = Link(
-    permissions=(permission_document_view,), text=_('Smart links'),
-    view='linking:smart_link_instances_for_document', args='object.pk'
+    icon='fa fa-link', permissions=(permission_document_view,),
+    text=_('Smart links'), view='linking:smart_link_instances_for_document',
+    args='object.pk'
 )
 link_smart_link_list = Link(
     permissions=(permission_smart_link_create,), text=_('Smart links'),