+++ mayan/apps/metadata/links.py
@@ -35,8 +35,8 @@ link_metadata_remove = Link(
     args='object.pk'
 )
 link_metadata_view = Link(
-    permissions=(permission_metadata_document_view,), text=_('Metadata'),
-    view='metadata:metadata_view', args='object.pk'
+    icon='fa fa-pencil', permissions=(permission_metadata_document_view,),
+    text=_('Metadata'), view='metadata:metadata_view', args='object.pk'
 )
 link_setup_document_type_metadata = Link(
     permissions=(permission_document_type_edit,), text=_('Optional metadata'),