+++ mayan/apps/ocr/links.py
@@ -10,8 +10,8 @@ from .permissions import (
 )
 
 link_document_content = Link(
-    permissions=(permission_ocr_content_view,), text=_('OCR'),
-    view='ocr:document_content', args='resolved_object.id'
+    icon='fa fa-font', permissions=(permission_ocr_content_view,),
+    text=_('OCR'), view='ocr:document_content', args='resolved_object.id'
 )
 link_document_submit = Link(
     permissions=(permission_ocr_document,), text=_('Submit for OCR'),